Refactor to have the actual checklists outside of EFB for later use in ECL
