+++ fbw-common/src/systems/instruments/src/EFB/Checklists/ChecklistsPage.tsx
@@ -4,8 +4,9 @@
 /* eslint-disable max-len */
 import React from 'react';
 import { ScrollableContainer } from '@flybywiresim/flypad';
+import { getAircraftChecklists } from '@flybywiresim/checklists';
 import { useAppSelector } from '../Store/store';
-import { ChecklistDefinition, getAircraftChecklists } from './Checklists';
+import { ChecklistDefinition } from './Checklists';
 import { ChecklistItemComponent } from './ChecklistItemComponent';
 import { CompletionButton } from './CompletionButton';
