+++ fbw-common/src/systems/shared/src/checklists/index.ts
@@ -1,0 +1,70 @@
+// Copyright (c) 2023-2024 FlyByWire Simulations
+// SPDX-License-Identifier: GPL-3.0
+
+import { getAircraftType } from '@flybywiresim/fbw-sdk';
+import {
+    afterLandingChecklistA32NX,
+    afterStartChecklistA32NX,
+    approachChecklistA32NX,
+    beforeStartChecklistA32NX,
+    cockpitPreparationChecklistA32NX,
+    landingChecklistA32NX,
+    lineUpChecklistA32NX,
+    parkingChecklistA32NX,
+    securingAircraftChecklistA32NX,
+    taxiChecklistA32NX,
+} from './Lists_A32NX/A32NX_Checklists';
+import { ChecklistDefinition } from '../../../instruments/src/EFB/Checklists/Checklists';
+import {
+    afterLandingChecklistA380X,
+    afterStartChecklistA380X,
+    afterTakeoffChecklistA380X,
+    approachChecklistA380X,
+    beforeStartChecklistA380X,
+    beforeTakeoffChecklistA380X,
+    landingChecklistA380X,
+    parkingChecklistA380X,
+    securingAircraftChecklistA380X,
+} from './Lists_A380X/A380X_Checklists';
+
+export const CHECKLISTS_A32NX: ChecklistDefinition[] = [
+    cockpitPreparationChecklistA32NX,
+    beforeStartChecklistA32NX,
+    afterStartChecklistA32NX,
+    taxiChecklistA32NX,
+    lineUpChecklistA32NX,
+    approachChecklistA32NX,
+    landingChecklistA32NX,
+    afterLandingChecklistA32NX,
+    parkingChecklistA32NX,
+    securingAircraftChecklistA32NX,
+];
+
+export const CHECKLISTS_A380X: ChecklistDefinition[] = [
+    beforeStartChecklistA380X,
+    afterStartChecklistA380X,
+    beforeTakeoffChecklistA380X,
+    afterTakeoffChecklistA380X,
+    approachChecklistA380X,
+    landingChecklistA380X,
+    afterLandingChecklistA380X,
+    parkingChecklistA380X,
+    securingAircraftChecklistA380X,
+];
+
+/**
+ * @brief Get the checklists for the current aircraft (A32NX or A380X)
+ *
+ * FIXME: use the correct getAircraftType function once PR #8500 is merged
+ */
+export function getAircraftChecklists() {
+    switch (getAircraftType()) {
+    case 'a32nx':
+        return CHECKLISTS_A32NX;
+    case 'a380x':
+        return CHECKLISTS_A380X;
+    default:
+        console.error('Unknown aircraft type for checklists');
+        return [];
+    }
+}
